# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mooda API",
        "description": "REST API Mooda untuk aplikasi mobile (Flutter). Semua endpoint terproteksi memakai Bearer token (Sanctum).",
        "contact": {
            "name": "Mooda Teknologi Indonesia",
            "email": "dev@mooda.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.mooda.id/api/v1",
            "description": "Produksi"
        },
        {
            "url": "http://127.0.0.1:8080/api/v1",
            "description": "Lokal (dev)"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login & terbitkan Bearer token",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "owner@mooda.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "mooda-mobile"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "3|abcd..."
                                                },
                                                "user": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kredensial tidak valid"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil pengguna saat ini",
                "operationId": "7edae2cdbaafc10dff8c9c4198b320a2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Belum login"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (hapus token saat ini)",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/fnb/inventory/ingredients": {
            "get": {
                "tags": [
                    "F&B - Inventory"
                ],
                "summary": "Daftar bahan + stok tersisa + status minimum",
                "operationId": "4ae94d5ad277bee55084638bc705fdf6",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/fnb/inventory/movements": {
            "post": {
                "tags": [
                    "F&B - Inventory"
                ],
                "summary": "Catat stok masuk/keluar (keluar termasuk waste)",
                "operationId": "2e487e3494fc6a74e3177113b1c7b3cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ingredient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "in",
                                        "enum": [
                                            "in",
                                            "out"
                                        ]
                                    },
                                    "qty": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "buy_price": {
                                        "description": "wajib utk type=in (harga/lot batch)",
                                        "type": "integer",
                                        "example": 13000,
                                        "nullable": true
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "pembelian / waste / rusak",
                                        "nullable": true
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tercatat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/fnb/inventory/opname": {
            "post": {
                "tags": [
                    "F&B - Inventory"
                ],
                "summary": "Stok opname (sesuaikan stok fisik)",
                "operationId": "9b1b3aa48f5959d3b7058f7b4b3f1e69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "ingredient_id": {
                                                    "type": "integer"
                                                },
                                                "counted_qty": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}